refactor: consolidate build-custom-node scripts into one
- Merge build-custom-node.js and build-custom-node-ci.js
- Single script now adapts to CI environment automatically
- Detects CI via CI or GITHUB_ACTIONS env vars
- Supports both CLI args and env vars for configuration
- Includes all CI features: GitHub Actions outputs, ccache, build summaries
- Update CI workflows to use consolidated script
- Remove redundant build-custom-node-ci.js

This eliminates maintenance burden of keeping two scripts in sync
